# Application Programming Interface/BuildingBug.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BuildingBugAPI",
    "version": "v1"
  },
  "paths": {
    "/api/v{version}/Cliente/login": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "summary": "Logar um cliente.",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteLoginRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Cliente/new": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "summary": "Criar um cliente.",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateClienteRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Cliente/update": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "summary": "Atualizar um cliente.",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateClienteRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Cliente/delete/{id}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "summary": "Eliminar um cliente.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Cliente/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Obter um cliente.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Cliente/all": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Obter todos os clientes.",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CodigoPostal/new": {
      "post": {
        "tags": [
          "CodigoPostal"
        ],
        "summary": "Criar um código postal",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCodigoPostalRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CodigoPostal/update": {
      "put": {
        "tags": [
          "CodigoPostal"
        ],
        "summary": "Atualizar um código postal",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCodigoPostalRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CodigoPostal/delete/{cp}": {
      "delete": {
        "tags": [
          "CodigoPostal"
        ],
        "summary": "Eliminar um código postal",
        "parameters": [
          {
            "in": "path",
            "name": "cp",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CodigoPostal/{cp}": {
      "get": {
        "tags": [
          "CodigoPostal"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cp",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/CodigoPostal/all": {
      "get": {
        "tags": [
          "CodigoPostal"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/Servico/all/Cliente/{clienteId}": {
      "get": {
        "tags": [
          "Servico"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ClienteLoginRequestContract": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "CreateClienteRequestContract": {
      "required": [
        "apelido",
        "contacto",
        "email",
        "nome",
        "senha"
      ],
      "type": "object",
      "properties": {
        "nome": {
          "minLength": 1,
          "type": "string"
        },
        "apelido": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "senha": {
          "minLength": 1,
          "type": "string"
        },
        "contacto": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateCodigoPostalRequestContract": {
      "type": "object",
      "properties": {
        "cp": {
          "format": "int32",
          "type": "integer"
        },
        "distrito": {
          "type": "string"
        }
      }
    },
...
    "UpdateClienteRequestContract": {
      "required": [
        "apelido",
        "contacto",
        "nome"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "minLength": 1,
          "type": "string"
        },
        "apelido": {
          "minLength": 1,
          "type": "string"
        },
        "contacto": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateCodigoPostalRequestContract": {
      "type": "object",
      "properties": {
        "cp": {
          "format": "int32",
          "type": "integer"
        },
        "distrito": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insert Token as: 'Bearer your_token_here'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}